Adding mark Attendance Function in API
I added a face detection function which enables taking attendance through api and also update the sample data for it.

one more function for load face data so that we can save the face data or a bunch of users with just one request,
LoadFace which take all the photo in to add folder and add their encoding to the database.